moving JQFilterWithQuery method to utils
