+++ pkg/iac-providers/tfplan/v1/load-file.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 
 	"github.com/accurics/terrascan/pkg/iac-providers/output"
-	"github.com/itchyny/gojq"
+	"github.com/accurics/terrascan/pkg/utils"
 	"go.uber.org/zap"
 )
 
@@ -47,7 +47,7 @@ func (t *TFPlan) LoadIacFile(absFilePath string) (allResourcesConfig output.AllR
 	}
 
 	// run jq query on tfplan json
-	processed, err := JQFilterWithQuery(jqQuery, tfjson)
+	processed, err := utils.JQFilterWithQuery(jqQuery, tfjson)
 	if err != nil {
 		errMsg := fmt.Sprintf("failed to process tfplan JSON. error: '%v'", err)
 		zap.S().Debug(errMsg)
@@ -84,44 +84,4 @@ func getTFID(id string) string {
 		return strings.Join(split, ".")
 	}
 	return strings.Join(split[len(split)-2:], ".")
-}
-
-// JQFilterWithQuery runs jq query on the given input and returns the output
-func JQFilterWithQuery(jqQuery string, jsonInput []byte) ([]byte, error) {
-
-	var processed []byte
-
-	// convert read json input into map[string]interface{}
-	var input map[string]interface{}
-	if err := json.Unmarshal(jsonInput, &input); err != nil {
-		return processed, fmt.Errorf("failed to decode input JSON. error: '%v'", err)
-	}
-
-	// parse jq query
-	query, err := gojq.Parse(jqQuery)
-	if err != nil {
-		return processed, fmt.Errorf("failed to parse jq query. error: '%v'", err)
-	}
-
-	// run jq query on input
-	iter := query.Run(input)
-	for {
-		v, ok := iter.Next()
-		if !ok {
-			break
-		}
-		if err, ok := v.(error); ok {
-			zap.S().Warn("error in processing jq query; error: '%v'", err)
-			continue
-		}
-
-		jqout, err := json.Marshal(v)
-		if err != nil {
-			zap.S().Warn("failed to encode jq output into JSON. error: '%v'", err)
-			continue
-		}
-		processed = append(processed, jqout...)
-	}
-
-	return processed, nil
 }
